+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - LASRV007 - 03/05/2023 - Not see root process.
 - LASRV009 - 03/05/2023 - Put autoremove.
+- LASRV006 - 03/05/2023 - Not see password.
 
 ### Fixed
 - LASRVUCF - 29/04/2023 - Quit server file.